--- a/docs/docs/install/config/index.md
+++ b/docs/docs/install/config/index.md
@@ -16,6 +16,9 @@ proctree:
     cache:
         process: 8192
         thread: 4096
+    cache-ttl:
+        process: 60
+        thread: 60
 
 capabilities:
     bypass: false