--- a/examples/config/global_config.yaml
+++ b/examples/config/global_config.yaml
@@ -8,6 +8,9 @@ proctree:
     # cache:
     #     process: 8192
     #     thread: 4096
+    # cache-ttl:
+    #     process: 120
+    #     thread: 120
 
 capabilities:
     bypass: false