@rscampos 2024-08-23
double check the benchmark in my env, good improvement!

> GOOS=linux CC=clang GOARCH=arm64 CGO_CFLAGS="-I/vagrant/dist/libbpf -I/vagrant/3rdparty/libbpf/include/uapi" CGO_LDFLAGS="-L/vagrant/dist/libbpf/obj -lbpf " go test -benchmem -run=^$ -bench ^BenchmarkProcessTree$ github.com/aquasecurity/tracee/pkg/proctree -benchtime=1000000x -race

```
BenchmarkProcessTree/GetProcessByHash-Concurrency1-4             612.0 ns/op  309.2 ns/op  49.45%
BenchmarkProcessTree/GetProcessByHash-Concurrency2-4            1508 ns/op   947.5 ns/op  37.20%
BenchmarkProcessTree/GetProcessByHash-Concurrency4-4            3133 ns/op   1918 ns/op   38.77%
BenchmarkProcessTree/GetProcessByHash-Concurrency8-4            6332 ns/op   3646 ns/op   42.44%
BenchmarkProcessTree/GetOrCreateProcessByHash-Concurrency1-4     667.8 ns/op  491.9 ns/op  26.35%
BenchmarkProcessTree/GetOrCreateProcessByHash-Concurrency2-4    1694 ns/op   1377 ns/op   18.71%
BenchmarkProcessTree/GetOrCreateProcessByHash-Concurrency4-4    3396 ns/op   2707 ns/op   20.29%
BenchmarkProcessTree/GetOrCreateProcessByHash-Concurrency8-4    6842 ns/op   5419 ns/op   20.83%
BenchmarkProcessTree/GetThreadByHash-Concurrency1-4              667.1 ns/op  499.3 ns/op  25.14%
BenchmarkProcessTree/GetThreadByHash-Concurrency2-4             1673 ns/op   1381 ns/op   17.46%
BenchmarkProcessTree/GetThreadByHash-Concurrency4-4             3405 ns/op   2615 ns/op   23.21%
BenchmarkProcessTree/GetThreadByHash-Concurrency8-4             6844 ns/op   5317 ns/op   22.31%
BenchmarkProcessTree/GetOrCreateThreadByHash-Concurrency1-4      697.7 ns/op  522.8 ns/op  25.09%
BenchmarkProcessTree/GetOrCreateThreadByHash-Concurrency2-4     1694 ns/op   1377 ns/op   18.72%
BenchmarkProcessTree/GetOrCreateThreadByHash-Concurrency4-4     3506 ns/op   2860 ns/op   18.44%
BenchmarkProcessTree/GetOrCreateThreadByHash-Concurrency8-4     6809 ns/op   5406 ns/op   20.60%
```